requirements.txt
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
scipy==1.7.1
web3==5.24.0
